Move target/field logic to metabase-lib
