+++ frontend/src/metabase/parameters/utils/dashboards.unit.spec.js
@@ -8,14 +8,9 @@ import {
   hasMatchingParameters,
   getFilteringParameterValuesMap,
   getParameterValuesSearchKey,
-  getTargetField,
   getDashboardUiParameters,
 } from "metabase/parameters/utils/dashboards";
-import {
-  SAMPLE_DATABASE,
-  PRODUCTS,
-  metadata,
-} from "__support__/sample_database_fixture";
+import { PRODUCTS, metadata } from "__support__/sample_database_fixture";
 import Field from "metabase-lib/lib/metadata/Field";
 
 describe("metabase/parameters/utils/dashboards", () => {
@@ -588,37 +583,6 @@ describe("metabase/parameters/utils/dashboards", () => {
         }),
       ).toEqual(
         "dashboardId: 123, parameterId: 456, query: abc, filteringParameterValues: []",
-      );
-    });
-  });
-
-  describe("getTargetField", () => {
-    const target = ["dimension", ["field", PRODUCTS.CATEGORY.id, null]];
-
-    it("should return null when given a card without a `dataset_query`", () => {
-      const card = {
-        id: 1,
-      };
-
-      expect(getTargetField(target, card, metadata)).toBe(null);
-    });
-
-    it("should return the field that maps to the mapping target", () => {
-      const field = PRODUCTS.CATEGORY;
-
-      const card = {
-        id: 1,
-        dataset_query: {
-          type: "query",
-          database: SAMPLE_DATABASE.id,
-          query: {
-            "source-table": PRODUCTS.id,
-          },
-        },
-      };
-
-      expect(getTargetField(target, card, metadata)).toEqual(
-        expect.objectContaining({ id: field.id }),
       );
     });
   });
